requirements.txt
Django>=3.2,<4.0
djangorestframework
dj-rest-auth
djangorestframework-simplejwt
psycopg2-binary
gunicorn
